replaced memory
