+++ zkevm-circuits/src/circuit_tools.rs
@@ -6,4 +6,3 @@ pub mod cached_region;
 pub mod cell_manager;
 pub mod gadgets;
 pub mod memory;
-pub mod memory_;
